Comment replace_triggered_by in image.tf
The goal is to make these terraform scripts part of schematics. However,
schematics does not support Terraform 1.2.x. Hence this commit removes the
features that would demand Terraform 1.2.x, there by enabling this product
to be deployable via schematics.

When Schematics moves to 1.2.x in the future(Q3: 22 is tentative), we
can enable it back on.

Signed-off-by: Vishwanatha Subbanna <vishwa@linux.vnet.ibm.com>