Use Ubuntu:20.04 as the base OS for image mover
Ubuntu 22.04 uses newer algorithms on openssl, and hence terraform 1.1.9
won't be able to connect to it. This commit changes the Ubuntu version
to be 20.04. The change also enables this repo to be used from inside
the schematics since the Terraform version supported in schematics
is 1.1.9 at the moment.

Refer: https://github.com/hashicorp/terraform/issues/30134

Signed-off-by: Vishwanatha Subbanna <vishwa@linux.vnet.ibm.com>